Create yuvxyb-math crate
