+++ yuvxyb-math/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "yuvxyb-math"
+description = "Math abstractions for yuvxyb"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+documentation = "https://docs.rs/yuvxyb-math"
+homepage = "https://github.com/rust-av/yuvxyb"
+repository = "https://github.com/rust-av/yuvxyb"
+
+# Should be the same as yuvxyb
+rust-version = "1.64.0"
+
+[features]
+default = ["fastmath"]
+fastmath = []
