+++ yuvxyb-math/src/lib.rs
@@ -1,0 +1,43 @@
+#![deny(clippy::all)]
+#![warn(clippy::nursery)]
+#![warn(clippy::pedantic)]
+#![allow(clippy::cast_possible_truncation)]
+#![allow(clippy::inline_always)]
+#![allow(clippy::module_name_repetitions)]
+#![allow(clippy::similar_names)]
+#![allow(clippy::doc_markdown)]
+#![warn(clippy::clone_on_ref_ptr)]
+#![warn(clippy::create_dir)]
+#![warn(clippy::dbg_macro)]
+#![warn(clippy::default_numeric_fallback)]
+#![warn(clippy::exit)]
+#![warn(clippy::filetype_is_file)]
+#![warn(clippy::float_cmp_const)]
+#![warn(clippy::if_then_some_else_none)]
+#![warn(clippy::lossy_float_literal)]
+#![warn(clippy::map_err_ignore)]
+#![warn(clippy::mem_forget)]
+#![warn(clippy::mod_module_files)]
+#![warn(clippy::multiple_inherent_impl)]
+#![warn(clippy::pattern_type_mismatch)]
+#![warn(clippy::rc_buffer)]
+#![warn(clippy::rc_mutex)]
+#![warn(clippy::rest_pat_in_fully_bound_structs)]
+#![warn(clippy::same_name_method)]
+#![warn(clippy::str_to_string)]
+#![warn(clippy::string_to_string)]
+#![warn(clippy::undocumented_unsafe_blocks)]
+#![warn(clippy::unnecessary_self_imports)]
+#![warn(clippy::unneeded_field_pattern)]
+#![warn(clippy::use_debug)]
+#![warn(clippy::verbose_file_reads)]
+
+mod cbrtf;
+mod matrix;
+mod mul_add;
+mod pow_exp;
+
+pub use cbrtf::cbrtf;
+pub use matrix::{ColVector, Matrix, RowVector};
+pub use mul_add::multiply_add;
+pub use pow_exp::{expf, powf};
